+++ package.json
@@ -7,7 +7,7 @@
     "node": "^8.0.0"
   },
   "scripts": {
-    "test-integration": "cross-env SECRETORKEY=kick-opossum-snowiness ISSUER=auth.reader-api.test DOMAIN=https://reader-api.test NODE_ENV=test tap -c -R classic --cov tests/integration/index.js",
+    "test-integration": "cross-env SECRETORKEY=kick-opossum-snowiness ISSUER=auth.reader-api.test DOMAIN=https://reader-api.test NODE_ENV=test tap -c --cov tests/integration/index.js",
     "test-google": "cross-env SECRETORKEY=kick-opossum-snowiness ISSUER=auth.reader-api.test DOMAIN=https://reader-api.test NODE_ENV=test tap -c -R classic --cov tests/google-integration/index.js",
     "test-models": "cross-env SECRETORKEY=kick-opossum-snowiness ISSUER=auth.reader-api.test DOMAIN=https://reader-api.test NODE_ENV=test tap -R classic --cov tests/models/index.js",
     "test-performance": "cross-env SECRETORKEY=kick-opossum-snowiness ISSUER=auth.reader-api.test DOMAIN=https://reader-api.test NODE_ENV=test tap -c -R min tests/performance/index.js",
@@ -53,7 +53,7 @@
     "sqlite3": "^4.0.4",
     "standard": "^11.0.1",
     "supertest": "^3.3.0",
-    "tap": "^12.6.3"
+    "tap": "^12.7.0"
   },
   "dependencies": {
     "@google-cloud/storage": "^2.3.1",
