+++ tests/integration/activity-create.test.js
@@ -1,16 +1,62 @@
 const request = require('supertest')
 const tap = require('tap')
 const urlparse = require('url').parse
-const { getToken, createUser, destroyDB } = require('../utils/utils')
+const {
+  getToken,
+  createUser,
+  destroyDB,
+  getActivityFromUrl,
+  createPublication
+} = require('../utils/utils')
+const { Document } = require('../../models/Document')
+const { Reader } = require('../../models/Reader')
+const { ReadActivity } = require('../../models/ReadActivity')
 const { urlToId } = require('../../utils/utils')
 
 const test = async app => {
   if (!process.env.POSTGRE_INSTANCE) {
     await app.initialize()
   }
+
   const token = getToken()
   const readerId = await createUser(app, token)
   const readerUrl = urlparse(readerId).path
+  let activityUrl
+
+  // Create Reader object
+  const person = {
+    name: 'J. Random Reader'
+  }
+  const reader1 = await Reader.createReader(readerId, person)
+
+  // Create Publication
+  const resActivity = await createPublication(app, token, readerUrl)
+
+  const activityUrl2 = resActivity.get('Location')
+  const activityObject = await getActivityFromUrl(app, activityUrl2, token)
+  const publicationUrl = activityObject.object.id
+
+  const resPublication = await request(app)
+    .get(urlparse(publicationUrl).path)
+    .set('Host', 'reader-api.test')
+    .set('Authorization', `Bearer ${token}`)
+    .type(
+      'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
+    )
+
+  // Create Document
+  const documentObject = {
+    mediaType: 'txt',
+    url: 'http://google-bucket/somewhere/file1234.txt',
+    documentPath: '/inside/the/book.txt',
+    json: { property1: 'value1' }
+  }
+
+  const document = await Document.createDocument(
+    reader1,
+    resPublication.body.id,
+    documentObject
+  )
 
   await tap.test('Create Activity', async () => {
     const res = await request(app)
@@ -219,117 +265,10 @@ const test = async app => {
     }
   )
 
-  await tap.test('Try to Delete something that is not valid', async () => {
-    const res = await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Delete',
-          object: {
-            type: 'SomethingInvalid'
-          }
-        })
-      )
-
-    await tap.equal(res.statusCode, 400)
-    const error = JSON.parse(res.text)
-    await tap.equal(error.statusCode, 400)
-    await tap.equal(error.error, 'Bad Request')
-    await tap.equal(error.details.badParams[0], 'object.type')
-    await tap.equal(error.details.type, 'SomethingInvalid')
-    await tap.equal(error.details.activity, 'Delete')
-  })
-
-  await tap.test('Try to Update something that is not valid', async () => {
-    const res = await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Update',
-          object: {
-            type: 'SomethingInvalid'
-          }
-        })
-      )
-
-    await tap.equal(res.statusCode, 400)
-    const error = JSON.parse(res.text)
-    await tap.equal(error.statusCode, 400)
-    await tap.equal(error.error, 'Bad Request')
-    await tap.equal(error.details.badParams[0], 'object.type')
-    await tap.equal(error.details.type, 'SomethingInvalid')
-    await tap.equal(error.details.activity, 'Update')
-  })
-
-  await tap.test('Get Outbox', async () => {
-    const res = await request(app)
-      .get(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-
-    await tap.equal(res.statusCode, 200)
-
-    const body = res.body
-    await tap.type(body, 'object')
-    await tap.type(body.id, 'string')
-    await tap.type(body['@context'], 'string')
-    await tap.equal(body.type, 'OrderedCollection')
-    await tap.type(body.summaryMap, 'object')
-    await tap.ok(Array.isArray(body.orderedItems))
-    await tap.type(body.orderedItems[0], 'object')
-    await tap.type(body.orderedItems[0].type, 'string')
-    // await tap.equal(activity.target, 'Publication') // No target inserted into table
-    await tap.equal(body.orderedItems[0].type, 'Create')
-    // await tap.type(body.orderedItems[0].actor, 'object')
-    // await tap.equal(body.orderedItems[0].actor.type, 'Person')
-    await tap.equal(urlToId(body.orderedItems[0].readerId), urlToId(readerId))
-    await tap.type(body.summaryMap, 'object')
-    await tap.type(body.orderedItems[0].id, 'string')
-  })
-
-  await tap.test('Get Outbox for reader that does not exist', async () => {
-    const res = await request(app)
-      .get(`${readerUrl}abc/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-
-    await tap.equal(res.statusCode, 404)
-    const error = JSON.parse(res.text)
-    await tap.equal(error.statusCode, 404)
-    await tap.equal(error.error, 'Not Found')
-    await tap.equal(error.details.type, 'Reader')
-    await tap.type(error.details.id, 'string')
-    await tap.equal(error.details.activity, 'Get Outbox')
-  })
-
+  await destroyDB(app)
   if (!process.env.POSTGRE_INSTANCE) {
     await app.terminate()
   }
-  await destroyDB(app)
 }
 
 module.exports = test
