+++ tests/integration/index.js
@@ -1,12 +1,38 @@
-const activityTests = require('./activity.test')
+const activityCreateTests = require('./activity-create.test')
+const activityGetTests = require('./activity-get.test')
+const readActivityCreateTests = require('./readActivity-create.test')
+
 const authErrorTests = require('./auth-error.test')
-const libraryTests = require('./library.test')
-const outboxTests = require('./outbox.test')
-const publicationTests = require('./publication.test')
-const readerTests = require('./reader.test')
-const noteTests = require('./note.test')
-const tagTests = require('./tag.test')
-const readerNotesTests = require('./reader-notes.test')
+
+const libraryGetTests = require('./library-get.test')
+const libraryPaginateTests = require('./library-paginate.test')
+const libraryFilterTests = require('./library-filter.test')
+const libraryOrderByTests = require('./library-orderBy.test')
+
+const noteCreateTests = require('./note-create.test')
+const noteGetTests = require('./note-get.test')
+const noteUpdateTests = require('./note-update.test')
+const noteDeleteTests = require('./note-delete.test')
+
+const outboxGetTests = require('./outbox-get.test')
+
+const publicationCreateTests = require('./publication-create.test')
+const publicationGetTests = require('./publication-get.test')
+const publicationUpdateTests = require('./publication-update.test')
+const publicationDeleteTests = require('./publication-delete.test')
+
+const readerCreateTests = require('./reader-create.test')
+const readerGetTests = require('./reader-get.test')
+
+const readerNotesGetTests = require('./readerNotes-get.test')
+const readerNotesPaginateTests = require('./readerNotes-paginate.test')
+const readerNotesFilterTests = require('./readerNotes-filter.test')
+const readerNotesOrderByTests = require('./readerNotes-orderBy.test')
+
+const tagCreateTests = require('./tag-create.test')
+const tagPublicationTests = require('./tag-publication.test')
+const tagNoteTests = require('./tag-note.test')
+const tagDeleteTests = require('./tag-delete.test')
 
 const app = require('../../server').app
 
@@ -23,15 +49,53 @@ const allTests = async () => {
 
   const test = process.env.npm_config_test
 
-  if (!test || test === 'activity') await activityTests(app)
+  if (!test || test === 'activity') {
+    await activityCreateTests(app)
+    await activityGetTests(app)
+    await readActivityCreateTests(app)
+  }
   if (!test || test === 'auth') await authErrorTests(app)
-  if (!test || test === 'library') await libraryTests(app)
-  if (!test || test === 'outbox') await outboxTests(app)
-  if (!test || test === 'publication') await publicationTests(app)
-  if (!test || test === 'reader') await readerTests(app)
-  if (!test || test === 'note') await noteTests(app)
-  if (!test || test === 'tag') await tagTests(app)
-  if (!test || test === 'readerNotes') await readerNotesTests(app)
+
+  if (!test || test === 'library') {
+    await libraryGetTests(app)
+    await libraryPaginateTests(app)
+    await libraryFilterTests(app)
+    await libraryOrderByTests(app)
+  }
+
+  if (!test || test === 'outbox') await outboxGetTests(app)
+
+  if (!test || test === 'publication') {
+    await publicationCreateTests(app)
+    await publicationGetTests(app)
+    await publicationUpdateTests(app)
+    await publicationDeleteTests(app)
+  }
+  if (!test || test === 'reader') {
+    await readerCreateTests(app)
+    await readerGetTests(app)
+  }
+
+  if (!test || test === 'note') {
+    await noteCreateTests(app)
+    await noteGetTests(app)
+    await noteUpdateTests(app)
+    await noteDeleteTests(app)
+  }
+
+  if (!test || test === 'tag') {
+    await tagCreateTests(app)
+    await tagPublicationTests(app)
+    await tagNoteTests(app)
+    await tagDeleteTests(app)
+  }
+
+  if (!test || test === 'readerNotes') {
+    await readerNotesGetTests(app)
+    await readerNotesPaginateTests(app)
+    await readerNotesFilterTests(app)
+    await readerNotesOrderByTests(app)
+  }
 
   if (process.env.POSTGRE_INSTANCE) {
     await app.knex.migrate.rollback()
