+++ tests/integration/note-create.test.js
@@ -1,0 +1,214 @@
+const request = require('supertest')
+const tap = require('tap')
+const urlparse = require('url').parse
+const {
+  getToken,
+  createUser,
+  destroyDB,
+  getActivityFromUrl,
+  createPublication
+} = require('../utils/utils')
+const { Document } = require('../../models/Document')
+const { Tag } = require('../../models/Tag')
+const { Note_Tag } = require('../../models/Note_Tag')
+const { Note } = require('../../models/Note')
+const { urlToId } = require('../../utils/utils')
+
+const test = async app => {
+  if (!process.env.POSTGRE_INSTANCE) {
+    await app.initialize()
+  }
+
+  const token = getToken()
+  const readerId = await createUser(app, token)
+  const readerUrl = urlparse(readerId).path
+  let noteUrl
+  let activityUrl
+
+  const resActivity = await createPublication(app, token, readerUrl)
+  const pubActivityUrl = resActivity.get('Location')
+  const pubActivityObject = await getActivityFromUrl(app, pubActivityUrl, token)
+  const publicationUrl = pubActivityObject.object.id
+
+  const resPublication = await request(app)
+    .get(urlparse(publicationUrl).path)
+    .set('Host', 'reader-api.test')
+    .set('Authorization', `Bearer ${token}`)
+    .type(
+      'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
+    )
+
+  // creating a document - this will not be exposed to the readers. It will be done as part of the upload
+  const createdDocument = await Document.createDocument(
+    { id: urlToId(readerId) },
+    urlToId(resPublication.body.id),
+    {
+      documentPath: '/path/1',
+      mediaType: 'text/html',
+      url: 'http://something/123'
+    }
+  )
+
+  const documentUrl = `${publicationUrl}${createdDocument.documentPath}`
+
+  await tap.test('Create Note', async () => {
+    const res = await request(app)
+      .post(`${readerUrl}/activity`)
+      .set('Host', 'reader-api.test')
+      .set('Authorization', `Bearer ${token}`)
+      .type(
+        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
+      )
+      .send(
+        JSON.stringify({
+          '@context': [
+            'https://www.w3.org/ns/activitystreams',
+            { reader: 'https://rebus.foundation/ns/reader' }
+          ],
+          type: 'Create',
+          object: {
+            type: 'Note',
+            content: 'This is the content of note A.',
+            'oa:hasSelector': { propety: 'value' },
+            context: publicationUrl,
+            inReplyTo: documentUrl,
+            noteType: 'test',
+            json: { property1: 'value1' }
+          }
+        })
+      )
+    await tap.equal(res.status, 201)
+    await tap.type(res.get('Location'), 'string')
+    activityUrl = res.get('Location')
+  })
+
+  await tap.test('Create Simple Note', async () => {
+    const res = await request(app)
+      .post(`${readerUrl}/activity`)
+      .set('Host', 'reader-api.test')
+      .set('Authorization', `Bearer ${token}`)
+      .type(
+        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
+      )
+      .send(
+        JSON.stringify({
+          '@context': [
+            'https://www.w3.org/ns/activitystreams',
+            { reader: 'https://rebus.foundation/ns/reader' }
+          ],
+          type: 'Create',
+          object: {
+            type: 'Note',
+            'oa:hasSelector': { propety: 'value' },
+            noteType: 'test'
+          }
+        })
+      )
+    await tap.equal(res.status, 201)
+    await tap.type(res.get('Location'), 'string')
+  })
+
+  await tap.test('Create Note with invalid document url', async () => {
+    const res = await request(app)
+      .post(`${readerUrl}/activity`)
+      .set('Host', 'reader-api.test')
+      .set('Authorization', `Bearer ${token}`)
+      .type(
+        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
+      )
+      .send(
+        JSON.stringify({
+          '@context': [
+            'https://www.w3.org/ns/activitystreams',
+            { reader: 'https://rebus.foundation/ns/reader' }
+          ],
+          type: 'Create',
+          object: {
+            type: 'Note',
+            'oa:hasSelector': { propety: 'value' },
+            context: publicationUrl,
+            inReplyTo: documentUrl + 'abc',
+            noteType: 'test'
+          }
+        })
+      )
+    await tap.equal(res.status, 404)
+    const error = JSON.parse(res.text)
+    await tap.equal(error.statusCode, 404)
+    await tap.equal(error.error, 'Not Found')
+    await tap.equal(error.details.type, 'Document')
+    await tap.type(error.details.id, 'string')
+    await tap.equal(error.details.activity, 'Create Note')
+  })
+
+  // TODO: migrate
+  // await tap.test(
+  //   'Try to create Note with invalid Publication context',
+  //   async () => {
+  //     const res = await request(app)
+  //       .post(`${readerUrl}/activity`)
+  //       .set('Host', 'reader-api.test')
+  //       .set('Authorization', `Bearer ${token}`)
+  //       .type(
+  //         'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
+  //       )
+  //       .send(
+  //         JSON.stringify({
+  //           '@context': [
+  //             'https://www.w3.org/ns/activitystreams',
+  //             { reader: 'https://rebus.foundation/ns/reader' }
+  //           ],
+  //           type: 'Create',
+  //           object: {
+  //             type: 'Note',
+  //             content: 'This is the content of note A.',
+  //             'oa:hasSelector': {},
+  //             context: publicationUrl + '123',
+  //             inReplyTo: documentUrl
+  //           }
+  //         })
+  //       )
+
+  //     await tap.equal(res.status, 404)
+  //   }
+  // )
+
+  // TODO: migrate
+  // await tap.test(
+  //   'Try to create Note with invalid inReplyTo Document',
+  //   async () => {
+  //     const res = await request(app)
+  //       .post(`${readerUrl}/activity`)
+  //       .set('Host', 'reader-api.test')
+  //       .set('Authorization', `Bearer ${token}`)
+  //       .type(
+  //         'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
+  //       )
+  //       .send(
+  //         JSON.stringify({
+  //           '@context': [
+  //             'https://www.w3.org/ns/activitystreams',
+  //             { reader: 'https://rebus.foundation/ns/reader' }
+  //           ],
+  //           type: 'Create',
+  //           object: {
+  //             type: 'Note',
+  //             content: 'This is the content of note A.',
+  //             'oa:hasSelector': {},
+  //             context: publicationUrl,
+  //             inReplyTo: documentUrl + '123'
+  //           }
+  //         })
+  //       )
+
+  //     await tap.equal(res.status, 404)
+  //   }
+  // )
+
+  if (!process.env.POSTGRE_INSTANCE) {
+    await app.terminate()
+  }
+  await destroyDB(app)
+}
+
+module.exports = test
