+++ tests/integration/publication-create.test.js
@@ -1,0 +1,168 @@
+const request = require('supertest')
+const tap = require('tap')
+const urlparse = require('url').parse
+const {
+  getToken,
+  createUser,
+  destroyDB,
+  getActivityFromUrl
+} = require('../utils/utils')
+const _ = require('lodash')
+const { urlToId } = require('../../utils/utils')
+const { Attribution } = require('../../models/Attribution')
+const { Document } = require('../../models/Document')
+const { Reader } = require('../../models/Reader')
+const { Tag } = require('../../models/Tag')
+const { Publication_Tag } = require('../../models/Publications_Tags')
+
+const test = async app => {
+  if (!process.env.POSTGRE_INSTANCE) {
+    await app.initialize()
+  }
+
+  const token = getToken()
+  const readerCompleteUrl = await createUser(app, token)
+  const readerUrl = urlparse(readerCompleteUrl).path
+
+  // Create Reader object
+  const person = {
+    name: 'J. Random Reader'
+  }
+
+  const reader1 = await Reader.createReader(readerCompleteUrl, person)
+
+  let publicationUrl
+  let activityUrl
+
+  const now = new Date().toISOString()
+
+  await tap.test('Create Publication', async () => {
+    const res = await request(app)
+      .post(`${readerUrl}/activity`)
+      .set('Host', 'reader-api.test')
+      .set('Authorization', `Bearer ${token}`)
+      .type(
+        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
+      )
+      .send(
+        JSON.stringify({
+          '@context': [
+            'https://www.w3.org/ns/activitystreams',
+            { reader: 'https://rebus.foundation/ns/reader' }
+          ],
+          type: 'Create',
+          object: {
+            type: 'Publication',
+            name: 'Publication A',
+            author: ['John Smith'],
+            editor: 'Jané S. Doe',
+            description: 'this is a description!!',
+            inLanguage: 'English',
+            datePublished: now,
+            links: [
+              {
+                '@context': 'https://www.w3.org/ns/activitystreams',
+                href: 'http://example.org/abc',
+                hreflang: 'en',
+                mediaType: 'text/html',
+                name: 'An example link'
+              }
+            ],
+            readingOrder: [
+              {
+                '@context': 'https://www.w3.org/ns/activitystreams',
+                href: 'http://example.org/abc',
+                hreflang: 'en',
+                mediaType: 'text/html',
+                name: 'An example reading order object1'
+              },
+              {
+                '@context': 'https://www.w3.org/ns/activitystreams',
+                href: 'http://example.org/abc',
+                hreflang: 'en',
+                mediaType: 'text/html',
+                name: 'An example reading order object2'
+              },
+              {
+                '@context': 'https://www.w3.org/ns/activitystreams',
+                href: 'http://example.org/abc',
+                hreflang: 'en',
+                mediaType: 'text/html',
+                name: 'An example reading order object3'
+              }
+            ],
+            resources: [
+              {
+                '@context': 'https://www.w3.org/ns/activitystreams',
+                href: 'http://example.org/abc',
+                hreflang: 'en',
+                mediaType: 'text/html',
+                name: 'An example resource'
+              }
+            ],
+            json: { property: 'value' }
+          }
+        })
+      )
+
+    await tap.equal(res.status, 201)
+    await tap.type(res.get('Location'), 'string')
+    activityUrl = res.get('Location')
+  })
+
+  await tap.test('Create Simple Publication', async () => {
+    const res = await request(app)
+      .post(`${readerUrl}/activity`)
+      .set('Host', 'reader-api.test')
+      .set('Authorization', `Bearer ${token}`)
+      .type(
+        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
+      )
+      .send(
+        JSON.stringify({
+          '@context': [
+            'https://www.w3.org/ns/activitystreams',
+            { reader: 'https://rebus.foundation/ns/reader' }
+          ],
+          type: 'Create',
+          object: {
+            type: 'Publication',
+            name: 'Publication A',
+            readingOrder: [
+              {
+                '@context': 'https://www.w3.org/ns/activitystreams',
+                href: 'http://example.org/abc',
+                hreflang: 'en',
+                mediaType: 'text/html',
+                name: 'An example reading order object2'
+              },
+              {
+                '@context': 'https://www.w3.org/ns/activitystreams',
+                href: 'http://example.org/abc',
+                hreflang: 'en',
+                mediaType: 'text/html',
+                name: 'An example reading order object'
+              },
+              {
+                '@context': 'https://www.w3.org/ns/activitystreams',
+                href: 'http://example.org/abc',
+                hreflang: 'en',
+                mediaType: 'text/html',
+                name: 'An example reading order object3'
+              }
+            ]
+          }
+        })
+      )
+
+    await tap.equal(res.status, 201)
+    await tap.type(res.get('Location'), 'string')
+  })
+
+  if (!process.env.POSTGRE_INSTANCE) {
+    await app.terminate()
+  }
+  await destroyDB(app)
+}
+
+module.exports = test
