+++ tests/integration/readActivity-create.test.js
@@ -21,7 +21,6 @@ const test = async app => {
   const token = getToken()
   const readerId = await createUser(app, token)
   const readerUrl = urlparse(readerId).path
-  let activityUrl
 
   // Create Reader object
   const person = {
@@ -57,86 +56,6 @@ const test = async app => {
     resPublication.body.id,
     documentObject
   )
-
-  await tap.test('Create Activity', async () => {
-    const res = await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Create',
-          object: {
-            type: 'Publication',
-            name: 'Publication A',
-            attributedTo: [
-              {
-                type: 'Person',
-                name: 'Sample Author'
-              }
-            ],
-            totalItems: 0,
-            orderedItems: []
-          }
-        })
-      )
-
-    await tap.equal(res.status, 201)
-    await tap.type(res.get('Location'), 'string')
-    activityUrl = res.get('Location')
-  })
-
-  // await tap.test('Get Activity', async () => {
-  //   const res = await request(app)
-  //     .get(urlparse(activityUrl).path)
-  //     .set('Host', 'reader-api.test')
-  //     .set('Authorization', `Bearer ${token}`)
-  //     .type(
-  //       'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-  //     )
-
-  //   await tap.equal(res.statusCode, 200)
-
-  //   const body = res.body
-
-  //   await tap.type(body, 'object')
-  //   await tap.type(body.id, 'string')
-  //   await tap.equal(body.type, 'Create')
-  //   await tap.type(body['@context'], 'object')
-  //   await tap.ok(Array.isArray(body['@context']))
-  //   await tap.equal(body.reader.id, readerId)
-  //   await tap.type(body.readerId, 'string')
-  //   await tap.type(body.object, 'object')
-  //   await tap.type(body.object.id, 'string')
-  //   await tap.type(body.reader.summaryMap.en, 'string')
-  //   await tap.type(body.actor, 'object')
-  //   await tap.type(body.actor.id, 'string')
-  //   await tap.equal(body.actor.type, 'Person')
-  // })
-
-  await tap.test('Get Activity that does not exist', async () => {
-    const res = await request(app)
-      .get(urlparse(activityUrl).path + 'abc')
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-    await tap.equal(res.statusCode, 404)
-    const error = JSON.parse(res.text)
-    await tap.equal(error.statusCode, 404)
-    await tap.equal(error.error, 'Not Found')
-    await tap.equal(error.details.type, 'Activity')
-    await tap.type(error.details.id, 'string')
-    await tap.equal(error.details.activity, 'Get Activity')
-  })
 
   await tap.test('Create Read activity with only a selector', async () => {
     const readActivity = await request(app)
