+++ tests/integration/tag-create.test.js
@@ -1,0 +1,146 @@
+const request = require('supertest')
+const tap = require('tap')
+const urlparse = require('url').parse
+const {
+  getToken,
+  createUser,
+  destroyDB,
+  getActivityFromUrl,
+  createPublication,
+  createNote
+} = require('../utils/utils')
+const { urlToId } = require('../../utils/utils')
+const { Document } = require('../../models/Document')
+const { Reader } = require('../../models/Reader')
+const { Note_Tag } = require('../../models/Note_Tag')
+const { Note } = require('../../models/Note')
+
+const test = async app => {
+  if (!process.env.POSTGRE_INSTANCE) {
+    await app.initialize()
+  }
+
+  const token = getToken()
+  const readerId = await createUser(app, token)
+  const readerUrl = urlparse(readerId).path
+  let stack
+
+  // Create Reader object
+  const person = {
+    name: 'J. Random Reader'
+  }
+  const reader1 = await Reader.createReader(readerId, person)
+
+  const resActivity = await createPublication(app, token, readerUrl)
+
+  const pubActivityUrl = resActivity.get('Location')
+  const pubActivityObject = await getActivityFromUrl(app, pubActivityUrl, token)
+  const publication = pubActivityObject.object
+
+  // Create a Document for that publication
+  const documentObject = {
+    mediaType: 'txt',
+    url: 'http://google-bucket/somewhere/file1234.txt',
+    documentPath: '/inside/the/book.txt',
+    json: { property1: 'value1' }
+  }
+  const document = await Document.createDocument(
+    reader1,
+    publication.id,
+    documentObject
+  )
+
+  const documentUrl = `${publication.id}${document.documentPath}`
+
+  // create Note for reader 1
+  const noteActivity = await createNote(app, token, readerUrl, {
+    inReplyTo: documentUrl,
+    context: publication.id
+  })
+
+  // get the urls needed for the tests
+  const noteActivityUrl = noteActivity.get('Location')
+
+  const noteActivityObject = await getActivityFromUrl(
+    app,
+    noteActivityUrl,
+    token
+  )
+  const noteUrl = noteActivityObject.object.id
+
+  await tap.test('Create Tag', async () => {
+    const res = await request(app)
+      .post(`${readerUrl}/activity`)
+      .set('Host', 'reader-api.test')
+      .set('Authorization', `Bearer ${token}`)
+      .type(
+        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
+      )
+      .send(
+        JSON.stringify({
+          '@context': [
+            'https://www.w3.org/ns/activitystreams',
+            { reader: 'https://rebus.foundation/ns/reader' }
+          ],
+          type: 'Create',
+          object: {
+            type: 'reader:Stack',
+            name: 'mystack',
+            json: { property: 'value' }
+          }
+        })
+      )
+    await tap.equal(res.status, 201)
+    await tap.type(res.get('Location'), 'string')
+    activityUrl = res.get('Location')
+  })
+
+  await tap.test('Try to create a duplicate Tag', async () => {
+    const res = await request(app)
+      .post(`${readerUrl}/activity`)
+      .set('Host', 'reader-api.test')
+      .set('Authorization', `Bearer ${token}`)
+      .type(
+        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
+      )
+      .send(
+        JSON.stringify({
+          '@context': [
+            'https://www.w3.org/ns/activitystreams',
+            { reader: 'https://rebus.foundation/ns/reader' }
+          ],
+          type: 'Create',
+          object: {
+            type: 'reader:Stack',
+            name: 'mystack'
+          }
+        })
+      )
+    await tap.equal(res.status, 400)
+    await tap.ok(res.error.text.startsWith('duplicate error:'))
+  })
+
+  await tap.test('Get tag when fetching library', async () => {
+    const res = await request(app)
+      .get(`${readerUrl}/library`)
+      .set('Host', 'reader-api.test')
+      .set('Authorization', `Bearer ${token}`)
+      .type(
+        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
+      )
+
+    await tap.equal(res.status, 200)
+    const body = res.body
+    await tap.ok(Array.isArray(body.tags))
+    await tap.type(body.tags[0].name, 'string')
+    await tap.type(body.tags[0].json, 'object')
+    stack = body.tags[0]
+  })
+
+  if (!process.env.POSTGRE_INSTANCE) {
+    await app.terminate()
+  }
+  await destroyDB(app)
+}
+
+module.exports = test
