--- tests/integration/tag.test.js
@@ -1,902 +1,0 @@
-const request = require('supertest')
-const tap = require('tap')
-const urlparse = require('url').parse
-const {
-  getToken,
-  createUser,
-  destroyDB,
-  getActivityFromUrl,
-  createPublication
-} = require('../utils/utils')
-const { urlToId } = require('../../utils/utils')
-const { Document } = require('../../models/Document')
-const { Reader } = require('../../models/Reader')
-const { Note_Tag } = require('../../models/Note_Tag')
-const { Note } = require('../../models/Note')
-
-const test = async app => {
-  if (!process.env.POSTGRE_INSTANCE) {
-    await app.initialize()
-  }
-
-  const token = getToken()
-  const readerId = await createUser(app, token)
-  const readerUrl = urlparse(readerId).path
-  let stack
-
-  // Create Reader object
-  const person = {
-    name: 'J. Random Reader'
-  }
-  const reader1 = await Reader.createReader(readerId, person)
-
-  const resActivity = await createPublication(app, token, readerUrl)
-
-  const pubActivityUrl = resActivity.get('Location')
-  const pubActivityObject = await getActivityFromUrl(app, pubActivityUrl, token)
-  const publication = pubActivityObject.object
-
-  // Create a Document for that publication
-  const documentObject = {
-    mediaType: 'txt',
-    url: 'http://google-bucket/somewhere/file1234.txt',
-    documentPath: '/inside/the/book.txt',
-    json: { property1: 'value1' }
-  }
-  const document = await Document.createDocument(
-    reader1,
-    publication.id,
-    documentObject
-  )
-
-  const documentUrl = `${publication.id}${document.documentPath}`
-
-  // create Note for reader 1
-  const noteActivity = await request(app)
-    .post(`${readerUrl}/activity`)
-    .set('Host', 'reader-api.test')
-    .set('Authorization', `Bearer ${token}`)
-    .type(
-      'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-    )
-    .send(
-      JSON.stringify({
-        '@context': [
-          'https://www.w3.org/ns/activitystreams',
-          { reader: 'https://rebus.foundation/ns/reader' }
-        ],
-        type: 'Create',
-        object: {
-          type: 'Note',
-          content: 'This is the content of note A.',
-          'oa:hasSelector': {},
-          context: publication.id,
-          inReplyTo: documentUrl,
-          noteType: 'test'
-        }
-      })
-    )
-
-  // get the urls needed for the tests
-  const noteActivityUrl = noteActivity.get('Location')
-
-  const noteActivityObject = await getActivityFromUrl(
-    app,
-    noteActivityUrl,
-    token
-  )
-  const noteUrl = noteActivityObject.object.id
-
-  await tap.test('Create Tag', async () => {
-    const res = await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Create',
-          object: {
-            type: 'reader:Stack',
-            name: 'mystack',
-            json: { property: 'value' }
-          }
-        })
-      )
-    await tap.equal(res.status, 201)
-    await tap.type(res.get('Location'), 'string')
-    activityUrl = res.get('Location')
-  })
-
-  await tap.test('Try to create a duplicate Tag', async () => {
-    const res = await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Create',
-          object: {
-            type: 'reader:Stack',
-            name: 'mystack'
-          }
-        })
-      )
-    await tap.equal(res.status, 400)
-    await tap.ok(res.error.text.startsWith('duplicate error:'))
-  })
-
-  await tap.test('Get tag when fetching library', async () => {
-    const res = await request(app)
-      .get(`${readerUrl}/library`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-
-    await tap.equal(res.status, 200)
-    const body = res.body
-    await tap.ok(Array.isArray(body.tags))
-    await tap.type(body.tags[0].name, 'string')
-    await tap.type(body.tags[0].json, 'object')
-    stack = body.tags[0]
-  })
-
-  await tap.test('Assign publication to tag', async () => {
-    const res = await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Add',
-          object: { id: stack.id, type: stack.type },
-          target: { id: urlToId(publication.id), type: 'Publication' }
-        })
-      )
-
-    await tap.equal(res.status, 201)
-
-    const pubres = await request(app)
-      .get(urlparse(publication.id).path)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-
-    await tap.equal(pubres.status, 200)
-    const body = pubres.body
-    await tap.ok(Array.isArray(body.tags))
-    await tap.equal(body.tags.length, 1)
-    await tap.equal(body.tags[0].type, 'reader:Stack')
-    await tap.equal(body.tags[0].name, 'mystack')
-  })
-
-  await tap.test(
-    'Try to assign publication to tag with invalid tag',
-    async () => {
-      const res = await request(app)
-        .post(`${readerUrl}/activity`)
-        .set('Host', 'reader-api.test')
-        .set('Authorization', `Bearer ${token}`)
-        .type(
-          'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-        )
-        .send(
-          JSON.stringify({
-            '@context': [
-              'https://www.w3.org/ns/activitystreams',
-              { reader: 'https://rebus.foundation/ns/reader' }
-            ],
-            type: 'Add',
-            object: { id: '999', type: stack.type },
-            target: { id: urlToId(publication.id), type: 'Publication' }
-          })
-        )
-      await tap.equal(res.status, 404)
-      const error = JSON.parse(res.text)
-      await tap.equal(error.statusCode, 404)
-      await tap.equal(error.details.type, 'Tag')
-      await tap.equal(error.details.activity, 'Add Tag to Publication')
-    }
-  )
-
-  await tap.test(
-    'Try to assign publication to tag with invalid publication',
-    async () => {
-      const res = await request(app)
-        .post(`${readerUrl}/activity`)
-        .set('Host', 'reader-api.test')
-        .set('Authorization', `Bearer ${token}`)
-        .type(
-          'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-        )
-        .send(
-          JSON.stringify({
-            '@context': [
-              'https://www.w3.org/ns/activitystreams',
-              { reader: 'https://rebus.foundation/ns/reader' }
-            ],
-            type: 'Add',
-            object: { id: stack.id, type: stack.type },
-            target: { id: 'notanid', type: 'Publication' }
-          })
-        )
-      await tap.equal(res.status, 404)
-      const error = JSON.parse(res.text)
-      await tap.equal(error.statusCode, 404)
-      await tap.equal(error.details.type, 'Publication')
-      await tap.equal(error.details.activity, 'Add Tag to Publication')
-    }
-  )
-
-  await tap.test('remove tag from publication', async () => {
-    const pubresbefore = await request(app)
-      .get(urlparse(publication.id).path)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-
-    await tap.equal(pubresbefore.status, 200)
-    const bodybefore = pubresbefore.body
-    await tap.ok(Array.isArray(bodybefore.tags))
-    await tap.equal(bodybefore.tags.length, 1)
-
-    const res = await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Remove',
-          object: stack,
-          target: { id: publication.id, type: 'Publication' }
-        })
-      )
-
-    await tap.equal(res.status, 201)
-
-    const pubres = await request(app)
-      .get(urlparse(publication.id).path)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-
-    await tap.equal(pubres.status, 200)
-    const body = pubres.body
-    await tap.ok(Array.isArray(body.tags))
-    await tap.equal(body.tags.length, 0)
-  })
-
-  await tap.test(
-    'Try to remove a tag from a publication with invalid tag',
-    async () => {
-      const res = await request(app)
-        .post(`${readerUrl}/activity`)
-        .set('Host', 'reader-api.test')
-        .set('Authorization', `Bearer ${token}`)
-        .type(
-          'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-        )
-        .send(
-          JSON.stringify({
-            '@context': [
-              'https://www.w3.org/ns/activitystreams',
-              { reader: 'https://rebus.foundation/ns/reader' }
-            ],
-            type: 'Remove',
-            object: { id: 12345, type: stack.type },
-            target: { id: publication.id, type: 'Publication' }
-          })
-        )
-      await tap.equal(res.status, 404)
-      const error = JSON.parse(res.text)
-      await tap.equal(error.statusCode, 404)
-      await tap.equal(error.details.type, 'Publication_Tag')
-      await tap.equal(error.details.activity, 'Remove Tag from Publication')
-    }
-  )
-
-  await tap.test(
-    'Try to remove a tag from a publication with invalid publication',
-    async () => {
-      const res = await request(app)
-        .post(`${readerUrl}/activity`)
-        .set('Host', 'reader-api.test')
-        .set('Authorization', `Bearer ${token}`)
-        .type(
-          'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-        )
-        .send(
-          JSON.stringify({
-            '@context': [
-              'https://www.w3.org/ns/activitystreams',
-              { reader: 'https://rebus.foundation/ns/reader' }
-            ],
-            type: 'Remove',
-            object: { id: stack.id, type: stack.type },
-            target: { id: 'notanid', type: 'Publication' }
-          })
-        )
-      await tap.equal(res.status, 404)
-      const error = JSON.parse(res.text)
-      await tap.equal(error.statusCode, 404)
-      await tap.equal(error.details.type, 'Publication_Tag')
-      await tap.equal(error.details.activity, 'Remove Tag from Publication')
-    }
-  )
-
-  await tap.test(
-    'Try to remove a tag from a publication with undefined tag',
-    async () => {
-      const res = await request(app)
-        .post(`${readerUrl}/activity`)
-        .set('Host', 'reader-api.test')
-        .set('Authorization', `Bearer ${token}`)
-        .type(
-          'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-        )
-        .send(
-          JSON.stringify({
-            '@context': [
-              'https://www.w3.org/ns/activitystreams',
-              { reader: 'https://rebus.foundation/ns/reader' }
-            ],
-            type: 'Remove',
-            object: { id: undefined, type: stack.type },
-            target: { id: publication.id, type: 'Publication' }
-          })
-        )
-      await tap.equal(res.status, 404)
-      const error = JSON.parse(res.text)
-      await tap.equal(error.statusCode, 404)
-      await tap.equal(error.details.type, 'Tag')
-      await tap.equal(error.details.activity, 'Remove Tag from Publication')
-    }
-  )
-
-  await tap.test(
-    'Try to remove a tag from a publication with undefined publication',
-    async () => {
-      const res = await request(app)
-        .post(`${readerUrl}/activity`)
-        .set('Host', 'reader-api.test')
-        .set('Authorization', `Bearer ${token}`)
-        .type(
-          'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-        )
-        .send(
-          JSON.stringify({
-            '@context': [
-              'https://www.w3.org/ns/activitystreams',
-              { reader: 'https://rebus.foundation/ns/reader' }
-            ],
-            type: 'Remove',
-            object: { id: stack.id, type: stack.type },
-            target: { id: undefined, type: 'Publication' }
-          })
-        )
-      await tap.equal(res.status, 404)
-      const error = JSON.parse(res.text)
-      await tap.equal(error.statusCode, 404)
-      await tap.equal(error.details.type, 'Publication')
-      await tap.equal(error.details.activity, 'Remove Tag from Publication')
-    }
-  )
-
-  // error disabled for now
-
-  await tap.test('Try to assign publication to tag twice', async () => {
-    await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Add',
-          object: stack,
-          target: { id: publication.id, type: 'Publication' }
-        })
-      )
-
-    const res = await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Add',
-          object: stack,
-          target: { id: publication.id, type: 'Publication' }
-        })
-      )
-    await tap.equal(res.status, 400)
-    const error = JSON.parse(res.text)
-    await tap.equal(error.statusCode, 400)
-    await tap.equal(error.details.type, 'Publication_Tag')
-    await tap.equal(error.details.activity, 'Add Tag to Publication')
-    await tap.type(error.details.target, 'string')
-    await tap.type(error.details.object, 'string')
-
-    /*
-    // doesn't affect the publication
-    const pubres = await request(app)
-      .get(urlparse(publication.id).path)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-
-    await tap.equal(pubres.status, 200)
-    const body = pubres.body
-    await tap.ok(Array.isArray(body.tags))
-    await tap.equal(body.tags.length, 1)
-    await tap.equal(body.tags[0].type, 'reader:Stack')
-    await tap.equal(body.tags[0].name, 'mystack')
-    */
-  })
-
-  await tap.test('Assign note to tag', async () => {
-    const res = await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Add',
-          object: { id: stack.id, type: stack.type },
-          target: { id: urlToId(noteUrl), type: 'Note' }
-        })
-      )
-
-    await tap.equal(res.status, 201)
-
-    const tagsForNotes = await request(app)
-      .get(urlparse(noteUrl).path)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-
-    await tap.equal(tagsForNotes.status, 200)
-    const body = tagsForNotes.body
-    await tap.ok(Array.isArray(body.tags))
-    await tap.equal(body.tags.length, 1)
-    await tap.equal(body.tags[0].type, 'reader:Stack')
-    await tap.equal(body.tags[0].name, 'mystack')
-  })
-
-  await tap.test('Try to assign note to tag with invalid tag', async () => {
-    const res = await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Add',
-          object: { id: 999, type: stack.type },
-          target: { id: urlToId(noteUrl), type: 'Note' }
-        })
-      )
-    await tap.equal(res.status, 404)
-    const error = JSON.parse(res.text)
-    await tap.equal(error.statusCode, 404)
-    await tap.equal(error.details.type, 'Tag')
-    await tap.equal(error.details.activity, 'Add Tag to Note')
-  })
-
-  await tap.test('Try to assign note to tag with invalid note', async () => {
-    const res = await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Add',
-          object: { id: stack.id, type: stack.type },
-          target: { id: 'notanid', type: 'Note' }
-        })
-      )
-    await tap.equal(res.status, 404)
-    const error = JSON.parse(res.text)
-    await tap.equal(error.statusCode, 404)
-    await tap.equal(error.details.type, 'Note')
-    await tap.equal(error.details.activity, 'Add Tag to Note')
-  })
-
-  await tap.test(
-    'try to assign a tag to an invalid type of object',
-    async () => {
-      const res = await request(app)
-        .post(`${readerUrl}/activity`)
-        .set('Host', 'reader-api.test')
-        .set('Authorization', `Bearer ${token}`)
-        .type(
-          'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-        )
-        .send(
-          JSON.stringify({
-            '@context': [
-              'https://www.w3.org/ns/activitystreams',
-              { reader: 'https://rebus.foundation/ns/reader' }
-            ],
-            type: 'Add',
-            object: { id: stack.id, type: 'SomethingInvalid' },
-            target: { id: 'notanid', type: 'Note' }
-          })
-        )
-      await tap.equal(res.status, 400)
-      const error = JSON.parse(res.text)
-      await tap.equal(error.statusCode, 400)
-      await tap.equal(error.details.badParams[0], 'object.type')
-      await tap.equal(error.details.activity, 'Add Tag to SomethingInvalid')
-    }
-  )
-
-  await tap.test('remove tag from note', async () => {
-    const note = await request(app)
-      .get(urlparse(noteUrl).path)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-
-    await tap.equal(note.status, 200)
-    const bodybefore = note.body
-    await tap.ok(Array.isArray(bodybefore.tags))
-    await tap.equal(bodybefore.tags.length, 1)
-
-    const res = await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Remove',
-          object: stack,
-          target: { id: urlToId(noteUrl), type: 'Note' }
-        })
-      )
-
-    await tap.equal(res.status, 201)
-
-    const notes = await request(app)
-      .get(urlparse(noteUrl).path)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-
-    await tap.equal(notes.status, 200)
-    const body = notes.body
-    await tap.ok(Array.isArray(body.tags))
-    await tap.equal(body.tags.length, 0)
-  })
-
-  await tap.test(
-    'Try to remove a tag from a note with invalid tag',
-    async () => {
-      const res = await request(app)
-        .post(`${readerUrl}/activity`)
-        .set('Host', 'reader-api.test')
-        .set('Authorization', `Bearer ${token}`)
-        .type(
-          'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-        )
-        .send(
-          JSON.stringify({
-            '@context': [
-              'https://www.w3.org/ns/activitystreams',
-              { reader: 'https://rebus.foundation/ns/reader' }
-            ],
-            type: 'Remove',
-            object: { id: 'blahTagId', type: stack.type },
-            target: { id: urlToId(noteUrl), type: 'Note' }
-          })
-        )
-      await tap.equal(res.status, 404)
-      const error = JSON.parse(res.text)
-      await tap.equal(error.statusCode, 404)
-      await tap.equal(error.details.type, 'Note_Tag')
-      await tap.equal(error.details.activity, 'Remove Tag from Note')
-    }
-  )
-
-  await tap.test(
-    'Try to remove a tag from a note with invalid note',
-    async () => {
-      const res = await request(app)
-        .post(`${readerUrl}/activity`)
-        .set('Host', 'reader-api.test')
-        .set('Authorization', `Bearer ${token}`)
-        .type(
-          'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-        )
-        .send(
-          JSON.stringify({
-            '@context': [
-              'https://www.w3.org/ns/activitystreams',
-              { reader: 'https://rebus.foundation/ns/reader' }
-            ],
-            type: 'Remove',
-            object: { id: stack.id, type: stack.type },
-            target: { id: 'notanid', type: 'Note' }
-          })
-        )
-      await tap.equal(res.status, 404)
-      const error = JSON.parse(res.text)
-      await tap.equal(error.statusCode, 404)
-      await tap.equal(error.details.type, 'Note_Tag')
-      await tap.equal(error.details.activity, 'Remove Tag from Note')
-    }
-  )
-
-  await tap.test(
-    'Try to remove a tag from a note with undefined note',
-    async () => {
-      const res = await request(app)
-        .post(`${readerUrl}/activity`)
-        .set('Host', 'reader-api.test')
-        .set('Authorization', `Bearer ${token}`)
-        .type(
-          'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-        )
-        .send(
-          JSON.stringify({
-            '@context': [
-              'https://www.w3.org/ns/activitystreams',
-              { reader: 'https://rebus.foundation/ns/reader' }
-            ],
-            type: 'Remove',
-            object: { id: stack.id, type: stack.type },
-            target: { id: undefined, type: 'Note' }
-          })
-        )
-      await tap.equal(res.status, 404)
-      const error = JSON.parse(res.text)
-      await tap.equal(error.statusCode, 404)
-      await tap.equal(error.details.type, 'Note')
-      await tap.equal(error.details.activity, 'Remove Tag from Note')
-    }
-  )
-
-  await tap.test(
-    'Try to remove a tag from a note with undefined tag',
-    async () => {
-      const res = await request(app)
-        .post(`${readerUrl}/activity`)
-        .set('Host', 'reader-api.test')
-        .set('Authorization', `Bearer ${token}`)
-        .type(
-          'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-        )
-        .send(
-          JSON.stringify({
-            '@context': [
-              'https://www.w3.org/ns/activitystreams',
-              { reader: 'https://rebus.foundation/ns/reader' }
-            ],
-            type: 'Remove',
-            object: { id: undefined, type: stack.type },
-            target: { id: urlToId(noteUrl), type: 'Note' }
-          })
-        )
-      await tap.equal(res.status, 404)
-      const error = JSON.parse(res.text)
-      await tap.equal(error.statusCode, 404)
-      await tap.equal(error.details.type, 'Tag')
-      await tap.equal(error.details.activity, 'Remove Tag from Note')
-    }
-  )
-
-  await tap.test(
-    'Try to delete a tag with a tagId that does not exist',
-    async () => {
-      const res = await request(app)
-        .post(`${readerUrl}/activity`)
-        .set('Host', 'reader-api.test')
-        .set('Authorization', `Bearer ${token}`)
-        .type(
-          'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-        )
-        .send(
-          JSON.stringify({
-            '@context': [
-              'https://www.w3.org/ns/activitystreams',
-              { reader: 'https://rebus.foundation/ns/reader' }
-            ],
-            type: 'Delete',
-            object: {
-              type: 'Tag',
-              id: stack.id + 'blah'
-            }
-          })
-        )
-
-      await tap.equal(res.statusCode, 404)
-      const error = JSON.parse(res.text)
-      await tap.equal(error.statusCode, 404)
-      await tap.equal(error.details.type, 'Tag')
-      await tap.equal(error.details.activity, 'Delete Tag')
-    }
-  )
-
-  await tap.test('Try to delete a tag with a null tagId', async () => {
-    const res = await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Delete',
-          object: {
-            type: 'Tag',
-            id: null
-          }
-        })
-      )
-
-    await tap.equal(res.statusCode, 404)
-    const error = JSON.parse(res.text)
-    await tap.equal(error.statusCode, 404)
-    await tap.equal(error.details.type, 'Tag')
-    await tap.equal(error.details.activity, 'Delete Tag')
-  })
-
-  await tap.test('Delete a tag', async () => {
-    // Get the library before the modifications
-    const libraryBefore = await request(app)
-      .get(`${readerUrl}/library`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-
-    // Add a tag to the note
-    await Note_Tag.addTagToNote(urlToId(noteUrl), libraryBefore.body.tags[0].id)
-
-    // Fetch the note with the tag
-    const noteWithTag = await Note.byId(urlToId(noteUrl))
-
-    await tap.equal(noteWithTag.tags.length, 1)
-    await tap.equal(noteWithTag.tags[0].name, libraryBefore.body.tags[0].name)
-    await tap.equal(libraryBefore.body.tags.length, 1)
-    await tap.equal(libraryBefore.body.tags[0].name, stack.name)
-    await tap.equal(libraryBefore.body.items[0].tags.length, 1)
-    await tap.equal(libraryBefore.body.items[0].tags[0].name, stack.name)
-
-    // Delete the tag
-    const res = await request(app)
-      .post(`${readerUrl}/activity`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-      .send(
-        JSON.stringify({
-          '@context': [
-            'https://www.w3.org/ns/activitystreams',
-            { reader: 'https://rebus.foundation/ns/reader' }
-          ],
-          type: 'Delete',
-          object: {
-            type: 'Tag',
-            id: stack.id
-          }
-        })
-      )
-
-    await tap.equal(res.statusCode, 204)
-
-    // Get the library after the modifications
-    const libraryAfter = await request(app)
-      .get(`${readerUrl}/library`)
-      .set('Host', 'reader-api.test')
-      .set('Authorization', `Bearer ${token}`)
-      .type(
-        'application/ld+json; profile="https://www.w3.org/ns/activitystreams"'
-      )
-
-    // Get the note after the modifications
-    const noteWithoutTag = await Note.byId(urlToId(noteUrl))
-
-    await tap.equal(libraryAfter.body.tags.length, 0)
-    await tap.equal(libraryAfter.body.items[0].tags.length, 0)
-    await tap.equal(noteWithoutTag.tags.length, 0)
-  })
-
-  if (!process.env.POSTGRE_INSTANCE) {
-    await app.terminate()
-  }
-  await destroyDB(app)
-}
-
-module.exports = test
